Added support for building & testing on Windows platforms.
